version finale de la fonction reneParle
