+++ app/rasp_speaking/talk.py
@@ -1,0 +1,21 @@
+
+import os
+
+def reneParle(text):
+    """Synthese vocale du texte saisi (text), par le Raspberry"""
+    parole = open("parole.txt", "w") #création d'un fichier text
+    parole.write("""#!/bin/bash
+pico2wave -l fr-FR -w temp.wav '""" + text +
+"""'
+amixer sset 'PCM' 95%
+aplay -q temp.wav
+rm temp.wav""")
+    parole.write("pico2wave -l fr-FR -w temp.wav ")
+    parole.write(text)
+    parole.close()
+    os.rename('parole.txt','parole.sh') #fichier text --> fichier shell
+    os.system('sh parole.sh') #execution du fichier shell
+    os.remove('parole.sh') # suppression du fichier shell
+
+
+#reneParle('Bonjour')
